+++ tests/testthat/test-reconstruct.R
@@ -350,7 +350,7 @@ test_that("new_R_ps() works as expected", {
     # Avoid by using only ECCs with "Final" and "Useful" as the Last.stage.
     dplyr::filter(Last.stage != IEATools::last_stages$services) %>%
     # Calculate the input-output matrices which are inputs to the new_R function.
-    calc_io_mats() %>%
+    calc_io_mats(direction = "downstream") %>%
     # Make an R_prime matrix that gives the same the resource inputs to the economy.
     # For testing purposes!
     dplyr::mutate(
@@ -387,14 +387,13 @@ test_that("new_R_ps() works as expected", {
     expect_true(matsbyname::equal_byname(newRsameasoldR$W_prime[[i]], newRsameasoldR$expected_W[[i]]))
   }
 
-
   doubleR <- UKEnergy2000mats %>%
     tidyr::spread(key = "matrix.name", value = "matrix") %>%
     # When Last.stage is "services", we get units problems.
     # Avoid by using only ECCs with "Final" and "Useful" as the Last.stage.
     dplyr::filter(Last.stage != IEATools::last_stages$services) %>%
     # Calculate the input-output matrices which are inputs to the new_R function.
-    calc_io_mats() %>%
+    calc_io_mats(direction = "downstream") %>%
     # Make an R_prime matrix that gives twice the resource inputs to the economy.
     dplyr::mutate(
       R_prime = matsbyname::hadamardproduct_byname(2, R)
@@ -429,6 +428,30 @@ test_that("new_R_ps() works as expected", {
     expect_equal(doubleR$W_prime[[i]], doubleR$expected_W[[i]])
   }
 
+  # Check some values when the new R matrix has 1's in it.
+  # These tests are in preparation for converting new_R_ps()
+  # to use new calc_io_mats(direction =- "downstream").
+  unitaryR <- UKEnergy2000mats %>%
+    tidyr::spread(key = "matrix.name", value = "matrix") %>%
+    # When Last.stage is "services", we get units problems.
+    # Avoid by using only ECCs with "Final" and "Useful" as the Last.stage.
+    dplyr::filter(Last.stage != IEATools::last_stages$services) %>%
+    dplyr::mutate(
+      R_prime = matsbyname::hadamardproduct_byname(2, R)
+    )
+  unitaryR$R_prime[[1]]["Resources - Crude", "Crude"] <- 1
+  unitaryR$R_prime[[1]]["Resources - NG", "NG"] <- 1
+  unitaryR <- unitaryR %>%
+    calc_io_mats(direction = "downstream") %>%
+    new_R_ps()
+  expect_equal(unitaryR$U_prime[[1]]["Crude - Dist.", "Oil refineries"], 0.940157103)
+  expect_equal(unitaryR$U_prime[[1]]["Elect - Grid", "Crude dist."], 0.0005812516)
+  expect_equal(unitaryR$V_prime[[1]]["Petrol dist.", "Petrol - Dist."], 0.53022566)
+  expect_equal(unitaryR$V_prime[[1]]["Oil refineries", "Diesel"], 0.41017456)
+  expect_equal(unitaryR$Y_prime[[1]]["Elect - Grid", "Residential"], 0.1395004)
+  expect_equal(unitaryR$Y_prime[[1]]["Petrol - Dist.", "Transport"], 0.5202214)
+
+
   # Test when the units on Products in U are not all same.
   # Under those conditions, we expect that U_prime, V_prime, and Y_prime are all NA.
   # Input units are not all same for the Last.stage = "services" cases.
@@ -436,7 +459,7 @@ test_that("new_R_ps() works as expected", {
   WithDiffUnits <- UKEnergy2000mats %>%
     tidyr::spread(key = "matrix.name", value = "matrix") %>%
     # Calculate the input-output matrices which are inputs to the new_R function.
-    calc_io_mats() %>%
+    calc_io_mats(direction = "downstream") %>%
     # Make an R_prime matrix that gives twice the resource inputs to the economy.
     dplyr::mutate(
       R_prime = matsbyname::hadamardproduct_byname(2, R)